Add files via upload
In this challenge, I will be using two separate APIs together to create an IP Address Tracking app. The main challenge is to build out this IP Address Tracker app and get it looking as close to the design as possible. To get the IP Address locations, I will be using the IP Geolocation API by IPify. So if you've got something you'd like to practice, feel free to give it a go.

Users should be able to:

1.View the optimal layout depending on their device's screen size.
2.See hover states for all interactive elements on the page.
3.See their own IP Address on the map on the initial page load.
4.Search for any IP addresses or domains and see the key information and location.

I do not have access to the Figma sketch so the design is not pixel perfect.